Getting rid of name collisions
Got the GL shaders to compile
Removed the copy of RenderOpenGL (we use the one in agg)
Refactoring VeldridGL
