+++ Veldrid/VertexFormats/VertexPositionColor.cs
@@ -30,7 +30,7 @@ either expressed or implied, of the FreeBSD Project.
 using System.Numerics;
 using Veldrid;
 
-namespace MatterHackers.Veldird.VertexFormats
+namespace MatterHackers.VeldirdProvider.VertexFormats
 {
 	public struct VertexPositionColor
 	{
